+++ modules/tools/record_analyzer/tools/perception_obstacle_sender.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 ###############################################################################
 # Copyright 2018 The Apollo Authors. All Rights Reserved.
@@ -19,8 +19,8 @@
 import time
 import argparse
 import google.protobuf.text_format as text_format
-from cyber_py import cyber
-from cyber_py import cyber_time
+from cyber.python.cyber_py3 import cyber
+from cyber.python.cyber_py3 import cyber_time
 from modules.perception.proto import perception_obstacle_pb2
 
 
@@ -29,7 +29,7 @@ def update(perception_obstacles):
     now = cyber_time.Time.now().to_sec()
     perception_obstacles.header.timestamp_sec = now
     perception_obstacles.header.lidar_timestamp = \
-        (long(now) - long(0.5)) * long(1e9)
+        (int(now) - int(0.5)) * int(1e9)
 
     for perception_obstacle in perception_obstacles.perception_obstacle:
         perception_obstacle.timestamp = now - 0.5
