+++ modules/tools/rosbag/body_sensation_evaluation.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 ###############################################################################
 # Copyright 2018 The Apollo Authors. All Rights Reserved.
@@ -26,8 +26,8 @@ import math
 import sys
 import time
 
-from cyber_py import cyber
-from cyber_py.record import RecordReader
+from cyber.python.cyber_py3 import cyber
+from cyber.python.cyber_py3.record import RecordReader
 from modules.canbus.proto import chassis_pb2
 from modules.canbus.proto.chassis_pb2 import Chassis
 from modules.localization.proto import localization_pb2
@@ -88,7 +88,7 @@ class BodySensationCalculator(object):
             if msg.topic == kChassisTopic:
                 chassis.ParseFromString(msg.message)
                 _t = msg.timestamp
-                t = long(str(_t)) * pow(10, -9)
+                t = int(str(_t)) * pow(10, -9)
                 cur_status = chassis.driving_mode
                 if mode["status"] != cur_status:
                     if mode["status"] != 'UNKNOW':
@@ -148,7 +148,7 @@ class BodySensationCalculator(object):
             if msg.topic == kLocalizationTopic:
                 localization.ParseFromString(msg.message)
                 _t = msg.timestamp
-                t = long(str(_t)) * pow(10, -9)
+                t = int(str(_t)) * pow(10, -9)
                 self.timestamp = t
                 diff_bump_time = t - self._last_bump_time
                 if diff_bump_time <= BUMP_TIME_THRESHOLD:
